+++ CHANGELOG.md
@@ -9,7 +9,7 @@
   - Job lifecycle hooks: both logical, systemic lifecycle (on activated, on executed, around execution) and local, usercode lifecycle (before, after, and around the `perform` method)
   - Call lifecycle hooks: before, after, and around every GRPC call
   - Each callback receives a lifecycle object which exposes the appropriate information (status, gauges, counters, timestamps, durations, configuration) for that lifecycle
-  - Callbacks may be prefiltered by attributes of the lifecycle object (e.g. register an after_job callback only for failed jobs)
+  - Callbacks may be prefiltered by attributes of the lifecycle object (e.g. register an after_perform callback only for failed jobs)
   - Lifecycle objects expose separate low-cardinality (for metric labels/tags) and high-cardinality (for logging) state projections
   - Support points for integrating with any APM / observability or error reporting platform: Datadog, NewRelic, Dynatrace, Airbrake, Sentry, etc.